requirements.txt
numpy
pandas
biopython
torch
rdkit==2023.3.2
joblib
tqdm
seaborn
molmap==1.3.9.8
mordred==1.2.0
mhfp==1.9.2
tmap==1.0.6
numba==0.56.4
scipy==1.7.3
rpy2==3.4.1
openpyxl
pytest
tomli; python_version < "3.11"
